Template tabs with alert for common templates
