+++ src/views/templates/details/tabs/disks/TemplateDisksPage.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router-dom';
 import produce from 'immer';
-import { isCommonTemplate } from 'src/views/clusteroverview/overview/components/inventory-card/utils/flattenTemplates';
 
 import { TemplateModel, V1Template } from '@kubevirt-ui/kubevirt-api/console';
 import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
@@ -18,6 +17,9 @@ import {
   useListPageFilter,
   VirtualizedTable,
 } from '@openshift-console/dynamic-plugin-sdk';
+
+import { isCommonVMTemplate } from '../../../utils';
+import NoEditableTemplateAlert from '../NoEditableTemplateAlert';
 
 import DiskListTitle from './components/DiskListTitle';
 import DiskRow from './components/DiskRow';
@@ -40,8 +42,7 @@ const TemplateDisksPage: React.FC<TemplateDisksPageProps> = ({ obj: template }) 
   const filters = useDisksFilters();
   const [data, filteredData, onFilterChange] = useListPageFilter(disks, filters);
   const vm = getTemplateVirtualMachineObject(template);
-
-  const isEditDisabled = isCommonTemplate(template);
+  const isEditDisabled = isCommonVMTemplate(template);
 
   const onUpdate = async (updatedVM: V1VirtualMachine) => {
     const updatedTemplate = produce(template, (draftTemplate) => {
@@ -58,6 +59,7 @@ const TemplateDisksPage: React.FC<TemplateDisksPageProps> = ({ obj: template }) 
 
   return (
     <div className="template-disk-page">
+      {isEditDisabled && <NoEditableTemplateAlert template={template} />}
       <ListPageHeader title="">
         <ListPageCreateButton
           isDisabled={isEditDisabled}
