+++ src/views/templates/details/tabs/parameters/TemplateParametersPage.tsx
@@ -7,6 +7,9 @@ import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTransla
 import { k8sPatch } from '@openshift-console/dynamic-plugin-sdk';
 import { ActionGroup, Alert, AlertVariant, Button, Divider, Title } from '@patternfly/react-core';
 
+import { isCommonVMTemplate } from '../../../utils';
+import NoEditableTemplateAlert from '../NoEditableTemplateAlert';
+
 import ParameterEditor from './ParameterEditor';
 
 import './template-parameters-page.scss';
@@ -20,13 +23,15 @@ type TemplateParametersPageProps = RouteComponentProps<{
 
 const TemplateParametersPage: React.FC<TemplateParametersPageProps> = ({ obj: template }) => {
   const { t } = useKubevirtTranslation();
+  const isEditDisabled = isCommonVMTemplate(template);
   const history = useHistory();
   const [parameters, updateParameters] = useImmer<TemplateParameter[]>(template.parameters);
   const [error, setError] = React.useState();
   const [success, setSuccess] = React.useState(false);
   const [loading, setLoading] = React.useState(false);
 
-  const isSaveDisabled = JSON.stringify(template.parameters) === JSON.stringify(parameters);
+  const isSaveDisabled =
+    JSON.stringify(template.parameters) === JSON.stringify(parameters) || isEditDisabled;
 
   const onParameterChange = (parameter: TemplateParameter) => {
     updateParameters((draftParameters) => {
@@ -62,53 +67,57 @@ const TemplateParametersPage: React.FC<TemplateParametersPageProps> = ({ obj: te
   };
 
   return (
-    <div className="co-m-pane__body template-parameters-page">
-      <Title headingLevel="h2" className="co-section-heading">
-        {t('Template parameters')}
-      </Title>
+    <>
+      {isEditDisabled && <NoEditableTemplateAlert template={template} />}
+      <div className="co-m-pane__body template-parameters-page">
+        <Title headingLevel="h2" className="co-section-heading">
+          {t('Template parameters')}
+        </Title>
 
-      <div className="co-m-pane__body co-m-pane__body--no-top-margin customize-vm">
-        <div className="row">
-          <div className="col-md-7">
-            {parameters.map((parameter, index) => (
-              <>
-                <ParameterEditor
-                  key={parameter.name}
-                  parameter={parameter}
-                  onChange={onParameterChange}
-                />
-                {index !== parameters.length - 1 && <Divider />}
-              </>
-            ))}
+        <div>
+          <div className="row">
+            <div className="col-md-7">
+              {parameters.map((parameter, index) => (
+                <>
+                  <ParameterEditor
+                    key={parameter.name}
+                    parameter={parameter}
+                    onChange={onParameterChange}
+                    isEditDisabled={isEditDisabled}
+                  />
+                  {index !== parameters.length - 1 && <Divider />}
+                </>
+              ))}
+            </div>
           </div>
         </div>
-      </div>
-      {error && (
-        <Alert variant={AlertVariant.danger} isInline title={t('Error')}>
-          {error}
-        </Alert>
-      )}
+        {error && (
+          <Alert variant={AlertVariant.danger} isInline title={t('Error')}>
+            {error}
+          </Alert>
+        )}
 
-      {success && (
-        <Alert variant={AlertVariant.info} isInline title={t('Success')}>
-          {t('Parameters successfully edited')}
-        </Alert>
-      )}
-      <ActionGroup className="pf-c-form">
-        <Button
-          isDisabled={isSaveDisabled}
-          type="submit"
-          variant="primary"
-          onClick={onSave}
-          isLoading={loading}
-        >
-          {t('Save')}
-        </Button>
-        <Button type="button" variant="secondary" onClick={goBack}>
-          {t('Cancel')}
-        </Button>
-      </ActionGroup>
-    </div>
+        {success && (
+          <Alert variant={AlertVariant.info} isInline title={t('Success')}>
+            {t('Parameters successfully edited')}
+          </Alert>
+        )}
+        <ActionGroup className="pf-c-form">
+          <Button
+            isDisabled={isSaveDisabled}
+            type="submit"
+            variant="primary"
+            onClick={onSave}
+            isLoading={loading}
+          >
+            {t('Save')}
+          </Button>
+          <Button type="button" variant="secondary" onClick={goBack}>
+            {t('Cancel')}
+          </Button>
+        </ActionGroup>
+      </div>
+    </>
   );
 };
