+++ src/views/templates/details/tabs/parameters/template-parameters-page.scss
@@ -1,4 +1,8 @@
 .template-parameters-page {
+  &.co-m-pane__body {
+    border-top: none;
+  }
+
   .pf-c-expandable-section__toggle-text {
     color: black;
   }
