+++ src/views/templates/details/tabs/yaml/TemplateYAMLPage.tsx
@@ -6,6 +6,9 @@ import Loading from '@kubevirt-utils/components/Loading/Loading';
 import { ResourceYAMLEditor } from '@openshift-console/dynamic-plugin-sdk';
 import { Bullseye } from '@patternfly/react-core';
 
+import { isCommonVMTemplate } from '../../../utils';
+import NoEditableTemplateAlert from '../NoEditableTemplateAlert';
+
 type TemplateYAMLPageProps = RouteComponentProps<{
   ns: string;
   name: string;
@@ -14,6 +17,7 @@ type TemplateYAMLPageProps = RouteComponentProps<{
 };
 
 const TemplateYAMLPage: React.FC<TemplateYAMLPageProps> = ({ obj: template }) => {
+  const isEditDisabled = isCommonVMTemplate(template);
   return (
     <React.Suspense
       fallback={
@@ -22,6 +26,7 @@ const TemplateYAMLPage: React.FC<TemplateYAMLPageProps> = ({ obj: template }) =>
         </Bullseye>
       }
     >
+      {isEditDisabled && <NoEditableTemplateAlert template={template} />}
       <ResourceYAMLEditor initialResource={template} />
     </React.Suspense>
   );
